requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
apscheduler==3.10.4
certifi==2023.11.17

# AI/ML
groq==0.4.0
transformers==4.36.0
torch==2.1.0
Pillow==10.1.0
numpy==1.26.0

# Document processing
PyMuPDF==1.23.8
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4